requirements.txt
fastapi
uvicorn
pydantic
ollama
PyMySQL
PyJWT
python-dotenv
